testing hella

+++ services/database/src/operations/delete-many.ts
@@ -1,9 +1,23 @@
+import * as E from 'express';
+import { BadRequestError, Utils } from 'recipiece-common';
+import { restoreObjectId } from '../utils';
 import { getCollection } from './util';
 
-export async function deleteManyOp(c: string, query: any): Promise<{deleted: number}> {
-  const collection = await getCollection(c);
-  const response = await collection.deleteMany(query);
-  return {
-    deleted: response.deletedCount,
-  };
+export async function deleteManyOp(req: E.Request, res: E.Response, next: E.NextFunction) {
+  const collectionName = req.params.collection;
+  const bodyQuery = req.body.query;
+  if (Utils.nou(bodyQuery)) {
+    next(new BadRequestError('query', 'undefined'));
+  } else {
+    const query = restoreObjectId(bodyQuery);
+    try {
+      const collection = await getCollection(collectionName);
+      const response = await collection.deleteMany(query);
+      res.status(200).send({
+        deleted: response.deletedCount,
+      });
+    } catch (e) {
+      next(e);
+    }
+  }
 }
